Do not use key-chord. Use as prefix M-; instead
It turns out that key-chord does not work when the input method is cyrillic (or
any other non-standard). M-; is the most convenient binding that we can afford
to change (commenting block). Use M-3 for commenting instead. Also, bind C-SPC
to yas-expand, since its previouis function is not that frequently used, and
can be easyly done with C-u M-SPC directly. It is not clear what will be the
best function for M-'. In this commit, it is set to run project-switch-project,
but it is not clear if this is a good idea.