Use M-; as a command prefix key
After lots of experiments, it seems that this is the optimal solution. Also,
this commit bring everything as close as possible to the "old" config. At the
end:

1. We have the M-; command prefix. From there commands are executed with 2
keys. For example "M-; d d" runs my-dired-at-home. There are also "express"
commands that are executed with "M-<letter>". For example "M-; M-s" runs
project-shell, "M-; M-m" runs magit-status, and so on.

2. Now M-; should be hit twise to (un)comment the region. This a minor
inconvenience.

3. Zap-to-char is bound to M-], not C-k. C-k is too valuable to waste on
something so rarely used.